big codebase refactor
