+++ src/hipp/kh9pc/fitting.py
@@ -1,0 +1,43 @@
+import numpy as np
+from numpy.typing import NDArray
+from sklearn.linear_model import LinearRegression, RANSACRegressor
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import PolynomialFeatures, StandardScaler
+
+
+def fit_ransac_poly(
+    x: NDArray[np.generic],
+    y: NDArray[np.generic],
+    degree: int = 3,
+    residual_threshold: float = 100,
+    max_trials: int = 100,
+) -> RANSACRegressor:
+    """Fit a polynomial regression with RANSAC on 1D data. Returns the fitted RANSACRegressor."""
+    poly_model = make_pipeline(
+        PolynomialFeatures(degree=degree),
+        StandardScaler(),
+        LinearRegression(),
+    )
+
+    min_samples = min(degree * 3, len(x))
+    ransac = RANSACRegressor(
+        poly_model, residual_threshold=residual_threshold, min_samples=min_samples, max_trials=max_trials
+    )
+    ransac.fit(x.reshape(-1, 1), y)
+    return ransac
+
+
+def detect_ruptures(vec: NDArray[np.number], threshold: float, reverse_scan: bool = False) -> NDArray[np.integer]:
+    """Detect indices where the signal drops below a threshold (falling edges).
+
+    If reverse_scan is True, scan from the end and return indices in original coordinates.
+    """
+    if reverse_scan:
+        vec = vec[::-1]
+
+    idx = np.where((vec[1:] <= threshold) & (vec[:-1] > threshold))[0] + 1
+
+    if reverse_scan:
+        idx = len(vec) - 1 - idx
+
+    return idx
