+++ src/hipp/kh9pc/qc/flat.py
@@ -1,0 +1,52 @@
+import matplotlib.pyplot as plt
+import rasterio
+from matplotlib.figure import Figure
+from rasterio.warp import Resampling
+from rasterio.windows import Window
+
+from hipp.kh9pc.restitution.flat import FlatStrategy
+
+
+def plot_flat_ruptures(detector: FlatStrategy) -> Figure:
+    """Band profiles (collapsed horizontally) with detected rupture row for top and bottom."""
+    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
+
+    for ax, side, result in zip(axes, ["top", "bottom"], [detector.top_, detector.bottom_]):
+        profile = result.sub_image.band.flatten()
+        ax.plot(profile, color="steelblue", linewidth=1)
+        ax.axvline(result.rupture_local, color="red", linewidth=1.5, label=f"rupture={result.rupture_local}")
+        ax.set_title(f"{side} band profile")
+        ax.set_xlabel("row index (downsampled)")
+        ax.set_ylabel("intensity")
+        ax.legend(fontsize=8)
+
+    return fig
+
+
+def plot_flat_edges(detector: FlatStrategy, margin_fraction: float = 0.03) -> Figure:
+    """Thumbnails around the top and bottom edge positions with detected line overlaid."""
+    left, _ = detector.vertical_detector.edges_
+    roi_w = detector.vertical_detector.detected_width_
+
+    fig, axes = plt.subplots(1, 2, figsize=(8, 4), constrained_layout=True)
+
+    with rasterio.open(detector.raster_filepath_) as src:
+        margin = int(margin_fraction * src.height)
+
+        for ax, side, result in zip(axes, ["top", "bottom"], [detector.top_, detector.bottom_]):
+            row_off = max(0, result.position - margin)
+            row_end = min(src.height, result.position + margin)
+            win_h = row_end - row_off
+            thumb = src.read(
+                1,
+                window=Window(left, row_off, roi_w, win_h),
+                out_shape=(512, 512),
+                resampling=Resampling.average,
+            )
+            line_row = (result.position - row_off) / win_h * 512
+            ax.imshow(thumb, cmap="gray", aspect="auto")
+            ax.axhline(line_row, color="yellow", linewidth=1.5)
+            ax.set_title(f"{side} edge — position={result.position} px")
+            ax.axis("off")
+
+    return fig
