+++ src/hipp/kh9pc/restitution/__init__.py
@@ -1,0 +1,7 @@
+from hipp.kh9pc.restitution.collimation import CollimationStrategy
+from hipp.kh9pc.restitution.fiducial import FiducialStrategy
+from hipp.kh9pc.restitution.flat import FlatStrategy
+from hipp.kh9pc.restitution.mixed import MixedStrategy
+from hipp.kh9pc.restitution.poly import PolyStrategy
+
+__all__ = ["CollimationStrategy", "FiducialStrategy", "FlatStrategy", "PolyStrategy", "MixedStrategy"]
