+++ src/hipp/kh9pc/restitution/mixed.py
@@ -2,11 +2,11 @@ import logging
 from dataclasses import dataclass, field
 from pathlib import Path
 
-from hipp.kh9pc.restitution_strategy.collimation_strategy import CollimationStrategy
-from hipp.kh9pc.restitution_strategy.fiducial_strategy import FiducialStrategy
-from hipp.kh9pc.restitution_strategy.flat_strategy import FlatStrategy
-from hipp.kh9pc.restitution_strategy.poly_strategy import PolyStrategy
-from hipp.kh9pc.types import RestitutionStrategy, Transformation
+from hipp.kh9pc.restitution.collimation import CollimationStrategy
+from hipp.kh9pc.restitution.fiducial import FiducialStrategy
+from hipp.kh9pc.restitution.flat import FlatStrategy
+from hipp.kh9pc.restitution.poly import PolyStrategy
+from hipp.kh9pc.restitution.base import RestitutionStrategy, Transformation
 
 logger = logging.getLogger(__name__)
