+++ src/hipp/kh9pc/restitution/templates/disk_15.png
