+++ src/hipp/kh9pc/restitution/templates/disk_20.png
