+++ src/hipp/kh9pc/restitution/templates/disk_25.png
