+++ src/hipp/kh9pc/restitution/templates/wagon_wheel.png
