+++ src/hipp/kh9pc/restitution/templates/weird_disk.png
