--- src/hipp/kh9pc/restitution_strategy/__init__.py
@@ -1,7 +1,0 @@
-from hipp.kh9pc.restitution_strategy.collimation_strategy import CollimationStrategy
-from hipp.kh9pc.restitution_strategy.fiducial_strategy import FiducialStrategy
-from hipp.kh9pc.restitution_strategy.flat_strategy import FlatStrategy
-from hipp.kh9pc.restitution_strategy.mixed_strategy import MixedStrategy
-from hipp.kh9pc.restitution_strategy.poly_strategy import PolyStrategy
-
-__all__ = ["CollimationStrategy", "FiducialStrategy", "FlatStrategy", "PolyStrategy", "MixedStrategy"]
